requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.106.0
gspread==6.2.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.7.0

# Additional utilities
click==8.1.7
